3214 - move Panel a level higher
